starting refactoring the layouts
