+++ app/root.tsx
@@ -2,15 +2,12 @@ import { type LinksFunction, type LoaderFunctionArgs } from "@remix-run/node";
 import { readUserSession } from "./utils/auth";
 
 import {
-  Link,
   Links,
   LiveReload,
   Meta,
-  NavLink,
   Outlet,
   Scripts,
   ScrollRestoration,
-  useLoaderData,
 } from "@remix-run/react";
 import stylesheet from "~/tailwind.css";
 
@@ -25,40 +22,7 @@ export async function loader({request}: LoaderFunctionArgs) {
   }
 }
 
-const LogInHeader = () => {
-  return (
-    <Link to="/login">Login</Link>
-  )
-}
-
-const LogOutHeader = () => {
-   return (<>
-      <form method="post" action="/logout" >
-        <button className="text-gray-200">Logout</button>
-      </form>
-                
-      <div className="flex flex-col items-end ">
-        <div className="text-md font-medium "></div>
-      </div>
-      <div className="h-10 w-10 rounded-full cursor-pointer bg-gray-200 border-2 border-blue-400"></div>
-      </>
-   )
-}
-
 export default function App() {
-
-  const { user }  = useLoaderData()
-  
-  // get user id
-  const getUserId = (user) => {
-    if(user === undefined){
-      return ''
-    } else {
-      return user.id
-    }
-  }
-  let id = getUserId(user)
-  
   return (
     <html lang="en">
       <head>
@@ -68,57 +32,7 @@ export default function App() {
         <Links />
       </head>
       <body>
-        <div className="h-screen w-full bg-white relative flex overflow-hidden">
-          <aside className="h-full w-16 pt-16 flex flex-col space-y-10 items-center justify-start relative bg-gray-800 text-white">
-            
-            {/* Icons https://flowbite.com/icons/  */}
-            {/* <!-- Goal Logger --> */}
-            <NavLink to="/">
-              <div className="h-10 w-10 flex items-center justify-center rounded-lg cursor-pointer hover:text-gray-800 hover:bg-white  hover:duration-300 hover:ease-linear focus:bg-white">
-                <svg className="w-6 h-6" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" width="18" height="20" fill="none" viewBox="0 0 18 20">
-                  <path stroke="currentColor" strokeLinecap="round" strokeWidth="2" d="M12 2h4a1 1 0 0 1 1 1v15a1 1 0 0 1-1 1H2a1 1 0 0 1-1-1V3a1 1 0 0 1 1-1h4m6 0v3H6V2m6 0a1 1 0 0 0-1-1H7a1 1 0 0 0-1 1M5 5h8m-5 5h5m-8 0h.01M5 14h.01M8 14h5"/>
-                </svg>
-              </div>
-            </NavLink>
-
-            {/* <!-- Calendar --> */}
-            {id === '' ? <div></div> : <NavLink to={`/calendar/2024/${id}`}>
-            <div className="h-10 w-10 flex items-center justify-center rounded-lg cursor-pointer hover:text-gray-800 hover:bg-white  hover:duration-300 hover:ease-linear focus:bg-white">
-              <svg className="w-6 h-6" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 20 20">
-                <path fill="currentColor" d="M6 1a1 1 0 0 0-2 0h2ZM4 4a1 1 0 0 0 2 0H4Zm7-3a1 1 0 1 0-2 0h2ZM9 4a1 1 0 1 0 2 0H9Zm7-3a1 1 0 1 0-2 0h2Zm-2 3a1 1 0 1 0 2 0h-2ZM1 6a1 1 0 0 0 0 2V6Zm18 2a1 1 0 1 0 0-2v2ZM5 11v-1H4v1h1Zm0 .01H4v1h1v-1Zm.01 0v1h1v-1h-1Zm0-.01h1v-1h-1v1ZM10 11v-1H9v1h1Zm0 .01H9v1h1v-1Zm.01 0v1h1v-1h-1Zm0-.01h1v-1h-1v1ZM10 15v-1H9v1h1Zm0 .01H9v1h1v-1Zm.01 0v1h1v-1h-1Zm0-.01h1v-1h-1v1ZM15 15v-1h-1v1h1Zm0 .01h-1v1h1v-1Zm.01 0v1h1v-1h-1Zm0-.01h1v-1h-1v1ZM15 11v-1h-1v1h1Zm0 .01h-1v1h1v-1Zm.01 0v1h1v-1h-1Zm0-.01h1v-1h-1v1ZM5 15v-1H4v1h1Zm0 .01H4v1h1v-1Zm.01 0v1h1v-1h-1Zm0-.01h1v-1h-1v1ZM2 4h16V2H2v2Zm16 0h2a2 2 0 0 0-2-2v2Zm0 0v14h2V4h-2Zm0 14v2a2 2 0 0 0 2-2h-2Zm0 0H2v2h16v-2ZM2 18H0a2 2 0 0 0 2 2v-2Zm0 0V4H0v14h2ZM2 4V2a2 2 0 0 0-2 2h2Zm2-3v3h2V1H4Zm5 0v3h2V1H9Zm5 0v3h2V1h-2ZM1 8h18V6H1v2Zm3 3v.01h2V11H4Zm1 1.01h.01v-2H5v2Zm1.01-1V11h-2v.01h2Zm-1-1.01H5v2h.01v-2ZM9 11v.01h2V11H9Zm1 1.01h.01v-2H10v2Zm1.01-1V11h-2v.01h2Zm-1-1.01H10v2h.01v-2ZM9 15v.01h2V15H9Zm1 1.01h.01v-2H10v2Zm1.01-1V15h-2v.01h2Zm-1-1.01H10v2h.01v-2ZM14 15v.01h2V15h-2Zm1 1.01h.01v-2H15v2Zm1.01-1V15h-2v.01h2Zm-1-1.01H15v2h.01v-2ZM14 11v.01h2V11h-2Zm1 1.01h.01v-2H15v2Zm1.01-1V11h-2v.01h2Zm-1-1.01H15v2h.01v-2ZM4 15v.01h2V15H4Zm1 1.01h.01v-2H5v2Zm1.01-1V15h-2v.01h2Zm-1-1.01H5v2h.01v-2Z"/>
-              </svg>
-            </div>
-            </NavLink>}
-
-            {/* <!-- List --> */}
-            {id === '' ? <div></div> : <NavLink to={`/list/2024/${id}`}>
-              <div className="h-10 w-10 flex items-center justify-center rounded-lg cursor-pointer hover:text-gray-800 hover:bg-white  hover:duration-300 hover:ease-linear focus:bg-white">
-                <svg className="w-6 h-6" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 17 10">
-                  <path stroke="currentColor" strokeLinecap="round" strokeWidth="2" d="M6 1h10M6 5h10M6 9h10M1.49 1h.01m-.01 4h.01m-.01 4h.01"/>
-                </svg>
-              </div>
-            </NavLink>}
-
-            {/* <!-- Dashboard --> */}
-            {id === '' ? <div></div> : <NavLink to={`/dashboard/2024/${id}`} >
-                <div className="h-10 w-10 flex items-center justify-center rounded-lg cursor-pointer hover:text-gray-800 hover:bg-white  hover:duration-300 hover:ease-linear focus:bg-white">
-                  <svg className="w-6 h-6" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 18 16">
-                    <path stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M1 1v14h16M4 10l3-4 4 4 5-5m0 0h-3.207M16 5v3.207"/>
-                  </svg>
-                </div>
-              </NavLink>}
-
-          </aside>
-          <div className="w-full h-full flex flex-col justify-between">
-            <header className="h-16 w-full flex items-center relative justify-end px-5 space-x-10 bg-gray-800">
-              <div className="flex flex-shrink-0 items-center space-x-4 text-white">
-                  {id === '' ? <LogInHeader />
-                  : <LogOutHeader />}
-              </div>
-            </header>
-            <Outlet context={id} />
-          </div>
-        </div>
+        <Outlet />
         <ScrollRestoration />
         <Scripts />
         <LiveReload />
