+++ app/routes/_auth.logout-success.tsx
@@ -1,10 +1,9 @@
 import type { MetaFunction } from "@remix-run/node";
-import { Link } from "@remix-run/react";
 
 export const meta: MetaFunction = () => {
   return [
-    { title: "365" },
-    { name: "description", content: "Welcome to Remix!" },
+    { title: "Logout Success" },
+    { name: "description", content: "You have been logged out" },
   ];
 };
 
@@ -13,7 +12,7 @@ export default function Index() {
 
   return (
     <main className="max-w-full h-full flex relative overflow-y-hidden bg-slate-100">
-      <div className="bg-green-500 w-full">test</div>
+      <div className="w-full text-center mt-8">Logout Successful</div>
   </main>
   );
 }
