+++ app/routes/_auth.tsx
@@ -1,0 +1,11 @@
+import { Outlet } from "@remix-run/react";
+
+export default function AuthLayout() {
+    return (
+        <>
+            <main>
+                <Outlet />
+            </main>
+        </>
+    )
+}
